moved login functions to login.py
